+ breakup core into modules
  - start of refactoring
