+++ tests/context.py
@@ -4,4 +4,4 @@ import sys
 import os
 sys.path.insert(0, os.path.abspath('..'))
 
-from svtyper import core
+from svtyper import core, statistics, utils, parsers
